Heating Gears Functionality
Added a microwave on the right. Uses a collision detection to detect when an object is placed in it. Then it checks if its the correct step and correct gear and if it is it starts a timer and ticking sound. I removed the tools on the right to fit the microwave here.